The code was reduced
