+++ app.js
@@ -9,6 +9,11 @@ var express     = require("express"),
     User        = require("./models/user"),
     seedDB      = require("./seeds");
 
+// Requiring routes
+var campgroundRoutes = require("./routes/campgrounds"),
+    commentRoutes    = require("./routes/comments"),
+    indexRoutes      = require("./routes/index");
+
 mongoose.set('useUnifiedTopology', true);
 mongoose.connect('mongodb://localhost:27017/yelp_camp', { useNewUrlParser: true });
 
@@ -38,139 +43,9 @@ app.use(function (req, res, next) {
     next();
 });
 
-app.get("/", function(req, res){
-    res.render("landing");
-});
-
-// Index - Show all campgrounds
-app.get("/campgrounds", function(req, res){
-    // Get all Campgrounds from DB
-    Campground.find({}, function(err, allCampgrounds) {
-        if(err){
-            console.log(err);
-        } else {
-            res.render("campgrounds/index", {campgrounds: allCampgrounds});
-        }
-    });
-});
-
-// Create - Add new campground to the database
-app.post("/campgrounds", function(req, res){
-    var name = req.body.name;
-    var image = req.body.image;
-    var desc = req.body.description;
-    var newCampground = {name: name, image: image, description: desc};
-    // Create new campground and save to DB
-    Campground.create(newCampground, function(err, newlyCreated) {
-        if(err){
-            console.log(err);
-        } else {
-            // Redirect to campgrounds page
-            res.redirect("/campgrounds");
-        }
-    });
-});
-
-// New - Create new campground form
-app.get("/campgrounds/new", function(req, res) {
-    res.render("campgrounds/new");
-});
-
-// Show - Show more info about the capground
-app.get("/campgrounds/:id", function(req, res) {
-    // Find campground with the provided ID
-    Campground.findById(req.params.id).populate("comments").exec(function(err, foundCampground) {
-        if(err) {
-            console.log(err);
-        } else {
-            // Render show template with that campground
-            res.render("campgrounds/show", {campground: foundCampground});
-        }
-    });
-});
-
-// ===============
-// Comments Route
-// ===============
-app.get("/campgrounds/:id/comments/new", isLoggedIn, function(req, res) {
-    Campground.findById(req.params.id, function(err, campground) {
-        if(err) {
-            console.log(err)
-        } else {
-            res.render("comments/new", {campground: campground});
-        }
-    });
-});
-
-app.post("/campgrounds/:id/comments", isLoggedIn, function(req, res) {
-    // lookup the campground using ID
-    Campground.findById(req.params.id, function(err, campground) {
-        if(err){
-            console.log(err);
-            res.redirect("/campgrounds");
-        } else {
-            // create a new comment
-            Comment.create(req.body.comment, function(err, comment) {
-                if(err){
-                    console.log(err);
-                } else {
-                    // reference comment to campground
-                    campground.comments.push(comment);
-                    campground.save();
-                    // redirect campground show page
-                    res.redirect("/campgrounds/" + campground._id);
-                }
-            });
-        }
-    });
-});
-
-//=====================
-// AUTH ROUTES
-//=====================
-
-// Show register form
-app.get("/register", function (req, res) {
-    res.render("register");
-});
-// Handle sign up logic
-app.post("/register", function (req, res) {
-    var newUser = new User({username: req.body.username});
-    User.register(newUser, req.body.password, function (err, user) {
-        if (err){
-            console.log(err);
-            return res.render("register");
-        }
-        passport.authenticate("local")(req, res, function () {
-            res.redirect("/campgrounds");
-        });
-    });
-});
-
-// Show login form
-app.get("/login", function (req, res) {
-    res.render("login");
-});
-// Handling login logic
-app.post("/login", passport.authenticate("local",
-    {
-        successRedirect: "/campgrounds",
-        failureRedirect: "/login"
-    }), function (req, res) {
-});
-
-// Logout route
-app.get("/logout", function (req, res) {
-    req.logout();
-    res.redirect("/campgrounds");
-});
-
-function isLoggedIn(req, res, next){
-    if (req.isAuthenticated()){
-        return next();
-    }
-    res.redirect("/login");
-}
+app.use("/", indexRoutes);
+app.use("/campgrounds", campgroundRoutes);
+app.use("/campgrounds/:id/comments", commentRoutes);
 
 app.listen(3000, process.env.IP, function(){
     console.log("The YelpCamp Server is listening...");
