+++ routes/comments.js
@@ -1,0 +1,50 @@
+var express = require("express");
+var router  = express.Router({mergeParams: true});
+var Comment = require("../models/comment");
+var Campground = require("../models/campground");
+
+// Comments new
+router.get("/new", isLoggedIn, function(req, res) {
+    Campground.findById(req.params.id, function(err, campground) {
+        if(err) {
+            console.log(err)
+        } else {
+            res.render("comments/new", {campground: campground});
+        }
+    });
+});
+
+
+// Comments create
+router.post("/", isLoggedIn, function(req, res) {
+    // lookup the campground using ID
+    Campground.findById(req.params.id, function(err, campground) {
+        if(err){
+            console.log(err);
+            res.redirect("/campgrounds");
+        } else {
+            // create a new comment
+            Comment.create(req.body.comment, function(err, comment) {
+                if(err){
+                    console.log(err);
+                } else {
+                    // reference comment to campground
+                    campground.comments.push(comment);
+                    campground.save();
+                    // redirect campground show page
+                    res.redirect("/campgrounds/" + campground._id);
+                }
+            });
+        }
+    });
+});
+
+// Middleware
+function isLoggedIn(req, res, next){
+    if (req.isAuthenticated()){
+        return next();
+    }
+    res.redirect("/login");
+}
+
+module.exports = router;
